requirements.txt
fastapi
uvicorn[standard]
elasticsearch
python-dotenv